refact: improve code structure
